# backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: FoodStock API
  description: Приложение для микромаркетов "на доверии"
  version: 1.0.0
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: "something went wrong"
    AdminAuthRequest:
      type: object
      properties:
        login:
          type: string
          example: "maket"
        password:
          type: string
          example: "maket123"
    AdminAuthResponse:
      type: object
      properties:
        token:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
...
    OrderItemRequest:
      type: object
      properties:
        item_id:
          type: string
          format: uuid
          example: "b3d538c5-3680-5bd1-b784-cebb32b29d6c"
        amount:
          type: integer
          description: "Количество единиц товара"
          example: 5
        price:
          type: integer
          format: int64
          description: "Цена в копейках за ОДНУ единицу товара. 10050 = 100 руб. 50 коп."
          example: 10050
    CreateOrderRequest:
      type: object
      properties:
        location_id:
          type: string
          format: uuid
          example: "b3d538c5-3680-5bd1-b784-cebb32b29d6c"
        items:
          type: array
          description: "Список товаров заказа"
          items:
            $ref: "#/components/schemas/OrderItemRequest"
    CreateOrderResponse:
      type: object
      properties:
        order_id:
          type: string
          format: uuid
          example: "b3d538c5-3680-5bd1-b784-cebb32b29d6c"
        total_price:
          type: integer
          format: int64
          description: "Общая сумма заказа в копейках. 10050 = 100 руб. 50 коп."
          example: 10050
        payment_url:
          type: string
          format: uri
          description: "Сгенерированная ссылка на оплату"
          example: "https://yoomoney.ru/checkout/payments/v2/contract?orderId=250270ad-000f-5000-8000-117600"
    LocationResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "b3d538c5-3680-5bd1-b784-cebb32b29d6c"
        slug:
          type: string
          example: "nsu_1"
        name:
          type: string
          example: "Холодильник в главном корпусе НГУ"
        address:
          type: string
          example: "Новосибирск, ул. Пирогова, д.1"
        is_active:
          type: boolean
          example: true
        created_at:
          type: string
          format: date-time
    CreateLocationRequest:
      type: object
      properties:
        slug:
          type: string
          example: "nsu_1"
        name:
          type: string
          example: "Холодильник в главном корпусе НГУ"
        address:
          type: string
          example: "Новосибирск, ул. Пирогова, д.1"
    CreateLocationResponse:
      type: object
      properties:
        location:
          $ref: "#/components/schemas/LocationResponse"
...
    ListLocationsResponse:
      type: object
      properties:
        locations:
          type: array
          items:
            $ref: "#/components/schemas/LocationResponse"
    NutritionResponse:
      type: object
      nullable: true
      description: "Калории и БЖУ"
      properties:
        calories:
          type: integer
          nullable: true
          example: 200
        proteins:
          type: number
          format: float
          nullable: true
          example: 23.6
        fats:
          type: number
          format: float
          nullable: true
          example: 1.9
        carbs:
          type: number
          format: float
          nullable: true
          example: 0.3
    ItemResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: "Сэндвич с курицей"
        description:
          type: string
          nullable: true
          example: "Сэндвич с курицей и соусом тар-тар"
        category:
          type: string
          example: "breakfast"
        photo_url:
          type: string
          example: "https://photos-storage/exsa129csa7690/chicken_sandwich.png"
        nutrition:
          $ref: "#/components/schemas/NutritionResponse"
        created_at:
          type: string
          format: date-time
    CreateItemRequest:
      type: object
      properties:
        name:
          type: string
          example: "Сэндвич с курицей"
        description:
          type: string
          nullable: true
          example: "Сэндвич с курицей и соусом тар-тар"
        category:
          type: string
          example: "breakfast"
        photo_url:
          type: string
          example: "https://photos-storage/exsa129csa7690/chicken_sandwich.png"
        nutrition:
          $ref: "#/components/schemas/NutritionResponse"
    CreateItemResponse:
      type: object
      properties:
        item:
          $ref: "#/components/schemas/ItemResponse"
...
    ListItemsResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ItemResponse"
...
  responses:
    BadRequestError:
      description: Ошибка в формате запроса
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: "invalid json"
    UnauthorizedError:
      description: Ошибка авторизации
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            missing_header:
              summary: Отсутствует заголовок
              value:
                error: "missing auth header"
            invalid_token:
              summary: Невалидный или просроченный токен
              value:
                error: "invalid or expired token"
...
    LocationIsNotOperationalError:
      description: Указанная торговая точка не принимает заказы
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: "location is not operational"
    InternalError:
      description: Внутренняя ошибка сервера
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: "internal error"
paths:
...
  /api/v1/client/orders:
    post:
      tags: [ Client ]
      summary: Создать заказ и получить ссылку на оплату
      description: Возвращает идентификатор заказа, итоговую сумму в копейках и ссылку на оплату
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '201':
          description: Заказ успешно создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateOrderResponse'
        '400':
          description: Некорректные данные
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                invalid_json:
                  summary: Битый JSON
                  value:
                    error: "invalid json"
                incorrect_location_id:
                  summary: Некорректный идентификатор торговой точки
                  value:
                    error: "invalid input"
                incorrect_items:
                  summary: Некорректный массив товаров (пустой или содержит невалидные объекты)
                  value:
                    error: "invalid input"
        '404':
          description: Ресурс не найден (торговая точка или конкретный товар)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                location_not_found:
                  summary: Торговая точка не найдена
                  value:
                    error: "location not found"
                location_item_not_found:
                  summary: Товар на точке не найден
                  value:
                    error: "location item not found"
        '409':
          description: Конфликт при создании заказа
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                order_exists:
                  summary: Заказ уже существует
                  value:
                    error: "order already exists"
                cannot_sell:
                  summary: Ошибка валидации товаров (цена/наличие)
                  description: "Возникает, если цена изменилась или товара нет в нужном количестве"
                  value:
                    error: "cannot sell one of the chosen items"
        '422':
          $ref: "#/components/responses/LocationIsNotOperationalError"
        '500':
          $ref: "#/components/responses/InternalError"
  /api/v1/admin/auth:
    post:
      tags: [ Admin Auth ]
      summary: Авторизация администратора
      description: Возвращает JWT токен доступа при успехе
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminAuthRequest'
      responses:
        '200':
          description: Токен успешно выдан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminAuthResponse'
        '400':
          $ref: "#/components/responses/BadRequestError"
        '401':
          description: Неверный логин или пароль
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "invalid login or password"
        '500':
          $ref: "#/components/responses/InternalError"
  /api/v1/admin/locations:
    post:
      tags: [ Admin Locations ]
      summary: Создать торговую точку
      description: Создаёт новую торговую точку и возвращает её данные
      security:
        - BearerAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateLocationRequest'
      responses:
        '201':
          description: Точка создана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateLocationResponse'
        '400':
          description: Некорректные данные
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                invalid_json:
                  summary: Битый JSON
                  value:
                    error: "invalid json"
                incorrect_slug:
                  summary: Некорректный slug
                  value:
                    error: "invalid input"
                incorrect_name:
                  summary: Некорректное имя
                  value:
                    error: "invalid input"
                incorrect_address:
                  summary: Некорректный адрес
                  value:
                    error: "invalid input"
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '409':
          description: Торговая точка уже существует
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "location with given slug already exists"
        '500':
          $ref: "#/components/responses/InternalError"
    get:
      tags: [ Admin Locations ]
      summary: Получить список всех торговых точек
      security:
        - BearerAuth: [ ]
      responses:
        '200':
          description: Список всех торговых точек
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListLocationsResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: "#/components/responses/InternalError"
...
  /api/v1/admin/items:
    post:
      tags: [ Admin Items ]
      summary: Создать товар
      description: Создаёт новый товар и возвращает его данные
      security:
        - BearerAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateItemRequest'
      responses:
        '201':
          description: Товар создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateItemResponse'
        '400':
          description: Некорректные данные
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                invalid_json:
                  summary: Битый JSON
                  value:
                    error: "invalid json"
                incorrect_name:
                  summary: Некорректное имя
                  value:
                    error: "invalid input"
                incorrect_description:
                  summary: Некорректное описание
                  value:
                    error: "invalid input"
                incorrect_category:
                  summary: Несуществующая категория
                  value:
                    error: "invalid input"
                incorrect_photo_url:
                  summary: Некорректная ссылка на фотографию
                  value:
                    error: "invalid input"
                incorrect_nutrition:
                  summary: Некорректное значение калорий или БЖУ
                  value:
                    error: "invalid input"
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: "#/components/responses/InternalError"
    get:
      tags: [ Admin Items ]
      summary: Получить список всех товаров
      security:
        - BearerAuth: [ ]
      responses:
        '200':
          description: Список всех товаров
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListItemsResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: "#/components/responses/InternalError"
...